init animate_cursor_eye
