+++ aopy/visualization/__init__.py
@@ -1,1 +1,2 @@
 from .base import *
+from .animation import *
